Move smoke test to separate file.
